Move feed.py to views/feed.py
